fixup target features. 现在应当能编译至 android 平台（比如在 Termux 上）。
