+++ Cargo.toml
@@ -18,6 +18,10 @@ name = "cxsign"
 version = "0.5.0"
 edition = "2021"
 
+[lib]
+name = "cxsign"
+crate-type = ["lib", "cdylib", "staticlib"]
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [features]
 default = ["completions"]
