+++ src/cli/mod.rs
@@ -55,6 +55,7 @@ fn match_signs(
         location_str,
         image,
         signcode,
+        #[cfg(any(target_os = "linux", target_os = "windows", target_os = "macos"))]
         precisely,
     } = cli_args;
     #[allow(clippy::mutable_key_type)]
@@ -76,6 +77,7 @@ fn match_signs(
                 location_str,
                 image,
                 &None,
+                #[cfg(any(target_os = "linux", target_os = "windows", target_os = "macos"))]
                 *precisely,
             )
             .sign(qs, sessions)?;
